+++ autoprotocol/liquid_handle/__init__.py
@@ -2,6 +2,7 @@
 Liquid handling utilities and methods utilized by Protocol.transfer to
 generate LiquidHandle instructions.
 """
+from .dispense import Dispense
 from .liquid_class import LiquidClass
 from .liquid_handle_method import LiquidHandleMethod
 from .mix import *