+++ docs/changelog.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+* :release:`7.8.0 <2021-05-18>`
+* :feature:`301` Add liquid_handle_dispense method, include protein buffer liquid class and volume resolution in `LiquidHandleBuilders`. Add helper function for building device level mode_params in LiquidHandleBuilders for tempest.
+* :support:`301` Add tests for liquid_handle_dispense method
+* :support:`301` Update authors
+
 * :release:`7.7.0 <2021-03-18>`
 * :feature:`297` Add support for SMILES to Compound class
 * :bug:`296 major` Remove low-evaporation cover type from FALCON96UBOTTOM